pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.4.0",
    "pytz>=2024.2",
    "requests>=2.32.3",
]
